+++ src/App.css
@@ -72,7 +72,7 @@ input[type="submit" i] {
 }
 
 footer {
-  position: relative;
+  position: absolute;
   bottom: 0px;
   background-color: #2a3d5f;
   min-height: 5vh;
