+++ src/App.js
@@ -1,7 +1,8 @@
 import React from "react";
 import "./App.css";
 import Posts from "./components/Posts";
-import Authentication from "./components/Authentication";
+import Register from "./components/Register";
+import Login from "./components/Login";
 import Route from "./components/Route";
 import Header from "./components/Header";
 
@@ -18,10 +19,17 @@ function App() {
           <div className="list_part">
             {/* when one jsx is provided another jsx tag the inner element is provided to the outer element as a prop called children , here the authentication component will show inside the route componet as a childern */}
             <Route path="/api/users/register">
-              <Authentication />
+              <Register />
             </Route>
           </div>
-          <div className="postlist_part">
+
+          <div className="list_part">
+            <Route path="/api/users/login">
+              <Login />
+            </Route>
+          </div>
+
+          <div className="list_part">
             <Route path="/api/posts">
               <Posts />
             </Route>
