+++ src/components/Register.js
@@ -1,14 +1,13 @@
 import React, { useState } from "react";
 import axios from "axios";
 
-function Authentication() {
+function Register() {
   //create states for differnt authentication parts to store the state of wht user is typing
   const [registerUserName, setRegisterUserName] = useState("");
   const [registerUserEmail, setRegisterUserEmail] = useState("");
   const [registerUserPassword, setRegisterUserPassword] = useState("");
-  const [loginUserEmail, setLoginUserEmail] = useState("");
-  const [loginUserPassword, setLoginUserPassword] = useState("");
-  const [data, setData] = useState(null);
+
+  // const [data, setData] = useState(null);
 
   const register = () => {
     axios({
@@ -30,33 +29,15 @@ function Authentication() {
       .catch();
   };
 
-  const login = () => {
-    axios({
-      method: "POST",
-      data: {
-        email: loginUserEmail,
-        password: loginUserPassword
-      },
-      withCredentials: false, //why?????
-      url: "https://syeda-recipe-api.herokuapp.com/api/users/login"
-    })
-      .then(res => {
-        console.log(res);
-        setLoginUserEmail("");
-        setLoginUserPassword("");
-      })
-      .catch();
-  };
-
-  const getUser = () => {
-    axios({
-      method: "GET",
-      withCredentials: false, //why?????
-      url: "https://syeda-recipe-api.herokuapp.com/api/users/all"
-    })
-      .then(res => setData(res.data))
-      .catch();
-  };
+  // const getUser = () => {
+  //   axios({
+  //     method: "GET",
+  //     withCredentials: false, //why?????
+  //     url: "https://syeda-recipe-api.herokuapp.com/api/users/all"
+  //   })
+  //     .then(res => setData(res.data))
+  //     .catch();
+  // };
 
   return (
     <div className="auth">
@@ -86,37 +67,16 @@ function Authentication() {
         </button>
       </div>
 
-      <div className="auth_parts">
-        <h2>Login</h2>
-        <label>
-          <input
-            placeholder="email"
-            onChange={e => setLoginUserEmail(e.target.value)}
-          />
-        </label>
-
-        <label>
-          <input
-            placeholder="password"
-            onChange={e => setLoginUserPassword(e.target.value)}
-          />
-        </label>
-
-        <button type="submit" onClick={login}>
-          Submit
-        </button>
-      </div>
-
-      <div className="auth_parts">
+      {/* <div className="auth_parts">
         <h2>Get User</h2>
 
         <button type="submit" onClick={getUser}>
           Submit
         </button>
         {data ? <h1>Welcome Back {data.name}</h1> : null}
-      </div>
+      </div> */}
     </div>
   );
 }
 
-export default Authentication;
+export default Register;
