+++ src/components/User.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+export default class User extends Component {
+  render() {
+    return (
+      <div className="onepost">
+        <h2 className="onepost_parts">`Name: ${this.props.user.name}`</h2>
+        <p className="onepost_parts">`Contact: ${this.props.user.email}`</p>
+        <p className="onepost_parts">`Member since: ${this.props.user.date}`</p>
+      </div>
+    );
+  }
+}
