+++ src/components/Users.js
@@ -1,0 +1,4 @@
+import React from "react";
+import axios from "axios";
+
+import User from "./User";
